Moved photos handling into photos service
